+++ Influx/Public/ConvertTo-Metric.ps1
@@ -20,6 +20,9 @@ Function ConvertTo-Metric {
         .PARAMETER TagProperty
             Optional: One or more strings which match property names of the Input Object, that you want to use as Tag data.
 
+        .PARAMETER TimeProperty
+            Optional: A string that matches a property name of the Input Object, that you want to use as the TimeStamp data.
+
         .PARAMETER Tags
             Optional: A hashtable of custom tag names and values.
 
@@ -47,11 +50,16 @@ Function ConvertTo-Metric {
         [String[]]
         $TagProperty,
 
+        [String]
+        $TimeProperty,
+
         [hashtable]
         $Tags
     )
     Process {
+
         ForEach ($ItemObject in $InputObject) {
+
             $Metrics = @{}
 
             ForEach ($Metric in $MetricProperty) {
@@ -75,12 +83,18 @@ Function ConvertTo-Metric {
 
             If ($Tags.Count -ne 0) { $TagData += $Tags }
 
-            [pscustomobject]@{
+            $Result = @{
                 PSTypeName = 'Metric'
                 Measure    = $Measure
                 Tags       = $TagData
                 Metrics    = $Metrics
             }
+
+            if ($ItemObject.$TimeProperty) {
+                $Result.add('TimeStamp', $(Get-Date $ItemObject.$TimeProperty))
+            }
+
+            [PSCustomObject]$Result
         }
     }
 }